Create gphg_selector.py
As yet incomplete script. The aim is to search through the full database of entrant watches (approx. 3000) and add a 1 to a column for those that were preselected (around 72 per year). I would also like to apply the same logic to then add a 1 in an additional column or columns for each watch that has won an individual prize (around 12 per year). I added the overall winners manually, since there are only 12.